Move DataTableHead to js/common/components
