+++ src/js/common/components/DataTable/index.js
@@ -1,0 +1,1 @@
+export { default as DataTableHead } from './DataTableHead';
